write custom UserCityAdapter class, implement equals() in City for comparison, change user city list style

+++ app/src/main/java/com/minghui_liu/android/lemonweather/database/UserCityDataSource.java
@@ -4,6 +4,12 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.widget.ArrayAdapter;
+
+import com.minghui_liu.android.lemonweather.model.City;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Created by kevin on 11/27/15.
@@ -12,6 +18,7 @@ public class UserCityDataSource {
     private static UserCityDataSource sUserCityDataSource;
     private Context mContext;
     private SQLiteDatabase mDatabase;
+    private ArrayList<City> mCityList;
 
     public static UserCityDataSource get(Context context) {
         if (sUserCityDataSource == null) {
@@ -20,35 +27,71 @@ public class UserCityDataSource {
         return sUserCityDataSource;
     }
 
-    private String[] allColumns = { UserCityContract.FeedEntry._ID,
-                                    UserCityContract.FeedEntry.COLUMN_NAME_CITY_NAME,
-                                    UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID };
-
     public UserCityDataSource(Context context) {
         mContext = context.getApplicationContext();
         mDatabase = new UserCityDbHelper(mContext).getWritableDatabase();
+        mCityList = new ArrayList<City>();
+        sync();
     }
 
-    public void addCity(String name, int id) {
-        ContentValues values = new ContentValues();
-        values.put(UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID, id);
-        values.put(UserCityContract.FeedEntry.COLUMN_NAME_CITY_NAME, name);
-        mDatabase.insert(UserCityContract.FeedEntry.TABLE_NAME, null, values);
+    public void addCity(City c) {
+        if (!mCityList.contains(c)) {
+            mCityList.add(c);
+            ContentValues values = new ContentValues();
+            values.put(UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID, c.getId());
+            values.put(UserCityContract.FeedEntry.COLUMN_NAME_CITY_NAME, c.getName());
+            mDatabase.insert(UserCityContract.FeedEntry.TABLE_NAME, null, values);
+        }
     }
 
-    public void removeCity(String name) {
-        String selection = UserCityContract.FeedEntry.COLUMN_NAME_CITY_NAME + " = ?";
-        String[] selectionArgs = { name };
-        mDatabase.delete(UserCityContract.FeedEntry.TABLE_NAME, selection, selectionArgs);
+    public void removeCity(City c) {
+        if (mCityList.contains(c)) {
+            mCityList.remove(c);
+            String selection = UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID + " = ?";
+            String[] selectionArgs = { String.valueOf(c.getId()) };
+            mDatabase.delete(UserCityContract.FeedEntry.TABLE_NAME, selection, selectionArgs);
+        }
     }
 
-    public void removeCity(int id) {
-        String selection = UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID + " = ?";
-        String[] selectionArgs = { String.valueOf(id) };
-        mDatabase.delete(UserCityContract.FeedEntry.TABLE_NAME, selection, selectionArgs);
+    /*public City getCity(int cityid) {
+        Cursor cursor = mDatabase.query(UserCityContract.FeedEntry.TABLE_NAME,
+                null,
+                UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID + " = ? ",
+                new String[] { String.valueOf(cityid) },
+                null,
+                null,
+                null
+                );
+        try {
+            if (cursor.getCount() == 0) {
+                return null;
+            }
+            cursor.moveToFirst();
+            String name = cursor.getString(cursor.getColumnIndex(UserCityContract.FeedEntry.COLUMN_NAME_CITY_NAME));
+            int id = cursor.getInt(cursor.getColumnIndex(UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID));
+            return new City(name, id);
+        } finally {
+            cursor.close();
+        }
+    }*/
+
+    public ArrayList<City> getAllCities() {
+        return mCityList;
     }
 
-    public Cursor getAllCities() {
-        return mDatabase.query(UserCityContract.FeedEntry.TABLE_NAME, allColumns, null, null, null, null, null);
+    public void sync() {
+        mCityList.clear();
+        Cursor cursor = mDatabase.query(UserCityContract.FeedEntry.TABLE_NAME, null, null, null, null, null, null);
+        try {
+            cursor.moveToFirst();
+            while (!cursor.isAfterLast()) {
+                String name = cursor.getString(cursor.getColumnIndex(UserCityContract.FeedEntry.COLUMN_NAME_CITY_NAME));
+                int id = cursor.getInt(cursor.getColumnIndex(UserCityContract.FeedEntry.COLUMN_NAME_CITY_ID));
+                mCityList.add(new City(name, id));
+                cursor.moveToNext();
+            }
+        } finally {
+            cursor.close();
+        }
     }
 }
